added payment gateway

+++ src/main/java/com/controller/OrderController.java
@@ -7,18 +7,25 @@ import javax.servlet.http.HttpServletRequest;
 import javax.transaction.Transactional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.servlet.ModelAndView;
+import org.springframework.web.servlet.View;
 
 import com.model.CartItemModel;
+import com.model.Order;
 import com.model.OrderModel;
 import com.model.UserModel;
+import com.paypal.api.payments.Links;
+import com.paypal.api.payments.Payment;
+import com.paypal.base.rest.PayPalRESTException;
 import com.repository.CartItemModelRepository;
 import com.repository.OrderModelRepository;
 import com.service.CartItemModelService;
 import com.service.OrderModelService;
+import com.service.PaypalService;
 import com.service.UserModelService;
 
 @Transactional
@@ -39,6 +46,12 @@ public class OrderController {
 	
 	@Autowired
 	OrderModelService orderModelService;
+	
+	@Autowired
+	PaypalService payPalService;
+
+	public static final String SUCCESS_URL = "pay/success";
+	public static final String CANCEL_URL = "pay/cancel";
 	
 	@GetMapping(path="orderDetails")
 	public ModelAndView orderDetails(ModelAndView mandv, HttpServletRequest request) {
@@ -69,20 +82,68 @@ public class OrderController {
         orderModel.setPaymentType(paymentType);
         orderModel.setUserId(userModel.getEmail());
         orderModel.setTotalPrice(cartItemModelService.getTotalPrice(userModel));
-        orderModel.setPaymentId("COD" + new Date().toString());
-        orderModel.setStatus("ordered");
         
-        orderModelRepository.save(orderModel);
+        orderModel.setStatus("yetToPay");
+   	 	orderModelRepository.save(orderModel);
         
-        System.out.println("ordermode id: " + orderModel.getId());
-        orderModelService.copyCartToOrders(userModel, orderModel);
+        if(paymentType.equals("OnlinePayment")) {
+        	
+        	try {
+
+           	 Order order = new Order(orderModel.getTotalPrice(), "USD", "PayPal", "sale", orderModel.getId().toString());
+          
+     			Payment payment = payPalService.createPayment(order.getPrice(), order.getCurrency(), order.getMethod(),
+     					order.getIntent(), order.getDescription(), "http://localhost:8080/" + CANCEL_URL,
+     					"http://localhost:8080/" + SUCCESS_URL);
+     			for(Links link:payment.getLinks()) {
+     				System.out.println("Link: " + link.getRel() + " : " + link.getHref());
+     				if(link.getRel().equals("approval_url")) {
+     					return new ModelAndView("redirect:"+link.getHref());
+     				}
+     			}
+     			
+     		} catch (PayPalRESTException e) {
+     		
+     			e.printStackTrace();
+     		}
+        	 
+        	 
+        	 
+        	 
+        }
+        else {
+        	orderModel.setPaymentId("COD" + new Date().toString());
+        	orderModel.setStatus("ordered");
+        	 
+        	orderModelRepository.save(orderModel);
+             
+            orderModelService.copyCartToOrders(userModel, orderModel);
+             
+            cartItemModelRepository.deleteAllByUserIdAndProceedToPayment(userModel, true);
+             
+            mandv.setViewName("redirect:/home");
+        }
         
-        cartItemModelRepository.deleteAllByUserIdAndProceedToPayment(userModel, true);
-        
-        mandv.setViewName("redirect:/home");
         return mandv;
 		
 	}
+	
+	@GetMapping(path="paymentSuccess")
+	public ModelAndView paymentSuccess(ModelAndView mandv, HttpServletRequest request) {
+        UserModel userModel = userModelService.extractUserModel(request);
+		
+        mandv.addObject("paymentStatus", "Payment Done Successfully! Order placed!");
+        
+        Set<OrderModel> Orders = orderModelRepository.findAllByUserId(userModel.getEmail());
+        mandv.addObject("orders", Orders);
+        
+        mandv.addObject("title", "Orders");
+        
+		mandv.setViewName("customer/orders");
+        return mandv;
+		
+	}
+	
 	
 	@GetMapping(path="orders")
 	public ModelAndView getUserOrders(ModelAndView mandv, HttpServletRequest request) {
